+++ packages/xchain-litecoin/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.12.1 (2023-05-18)
+
+## Add
+
+- New client function getAssetInfo() returns chain, decimals and asset
+- renamed defaultLTCParams to defaultLtcParams to be more consistent with CamelCase
+
 # v0.12.0 (2023-05-02)
 
 ## Update